Rewrite clement::router tests using Catch2
